% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{create_df}
\alias{create_df}
\title{Create and return a Pandas data frame}
\usage{
create_df()
}
\value{
a Pandas data frame
}
\description{
Dummy function to test using Pandas module from R.
}
\examples{
create_df()
}
