% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{install_pandas}
\alias{install_pandas}
\title{Install Python dependency}
\usage{
install_pandas(...)
}
\arguments{
\item{...}{any argument for \code{\link[reticulate:py_install]{reticulate::py_install()}}.}
}
\value{
no returned value.
}
\description{
Install the pandas package.
}
\examples{
\dontrun{
# create Python environement
venv <- "reticulate-test-env"
reticulate::virtualenv_create(venv)
# activate Python environment
reticulate::use_virtualenv(virtualenv = venv, required = TRUE)
# check version of Python
reticulate::py_config()
# install pandas
install_ktest(method = "virtualenv", envname = venv)
}
}
